Make tractor beam collision detection work smoother.
Previously, hitting the top of a rectangle while holding an item would
cause the player to jump. Now the position of the item in the beam is
adjusted a bit so as not to have to reset the player position. It's then
readjusted when vertically re-centered (such as during a jump).