Create a concept of "areas" per time period.
A TimePeriod is no longer a section of a level, but rather a container for
"areas" of a level. These can be multiple rooms, and switching time periods
within an area will attempt to go to the same time period in that area's
counterpart.

This is untested, beyond the default "outside" areas.